apps/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pylint==2.17.4
black==23.3.0
websockets==10.4
python-multipart==0.0.6
pydub==0.25.1
SpeechRecognition==3.10.0
gTTS==2.3.1
openai==0.27.8
manim==0.17.3
python-dotenv==1.0.0
